Add: Refactor note app to use Provider
