+++ fullstackopen-workshops/part-6/redux-notes/src/components/NoteForm.jsx
@@ -1,12 +1,31 @@
-import React from 'react'
+import React from "react";
+import { useSelector , useDispatch } from "react-redux";
+import { createNote} from "../reducers/noteReducer.js";
+const NoteForm = () => {
 
-const NoteForm = ({addNote}) => {
+  //state rw useState and this is the advantage of using redux
+  const dispatch = useDispatch();
+  const notes = useSelector((state) => state);
+
+  const addNote = (e) => {
+    e.preventDefault();
+    // console.dir(e.target);
+    console.dir(e.target.myInput.value);
+    const newNote = {
+      content: e.target.myInput.value,
+      important: true,
+      // id: store.getState().length + 1,
+      id: notes.length + 1,
+    };
+    dispatch(createNote(newNote));
+    e.target.myInput.value = ""; // clear the input field after adding a note
+  };
   return (
     <form onSubmit={addNote}>
-        <input name="myInput"/>
-        <button>Add Note</button>
+      <input name="myInput" />
+      <button>Add Note</button>
     </form>
-  )
-}
+  );
+};
 
-export default NoteForm
+export default NoteForm;
