+++ fullstackopen-workshops/part-6/redux-notes/src/components/Notes.jsx
@@ -1,0 +1,28 @@
+import React from 'react'
+import { useSelector, useDispatch } from 'react-redux';
+import { toggleImportanceOf } from "../reducers/noteReducer";
+
+
+const Notes = () => {
+    const dispatch = useDispatch();
+    const notes = useSelector((state) => state)
+
+    const toggleImportant = (id) => {
+    // e.preventDefault();
+    dispatch(toggleImportanceOf(id));
+  };
+  return (
+    <ul>
+        {notes.map((note) => (
+          <li key={note.id} onClick={() => {toggleImportant(note.id)}}>
+            {note.content}{" "}
+            <strong >
+              {note.important ? "important" : ""}
+            </strong>
+          </li>
+        ))}
+      </ul>
+  )
+}
+
+export default Notes
