+++ fullstackopen-workshops/part2/notes-app/.gitignore
@@ -7,6 +7,8 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
+
+coverage
 node_modules
 dist
 dist-ssr
