+++ fullstackopen-workshops/part3/notes-server/.env
@@ -1,0 +1,5 @@
+MONGODB_URI="mongodb+srv://tolangeakash753:mypassword@cluster0.rfc1bk0.mongodb.net/notesApp?retryWrites=true&w=majority&appName=Cluster0"
+TEST_MONGODB_URI="mongodb+srv://tolangeakash753:mypassword@cluster0.rfc1bk0.mongodb.net/testingdb?retryWrites=true&w=majority&appName=Cluster0"
+PORT=3005
+
+SECRET="mysecret pass"
